Added PipeSize as hard-coded parameter
The default PIPE_SIZE is 1kb. Which is results in way too many
wait-states/Thread-switches when processing a large volume of data
Code has been slightly restructured to accept a larger PipeSize
and the new PipeSize is set to 32kb hardcode.

Test-results have been published oni June 23 to clojure-dev:
https://groups.google.com/forum/#!topic/clojure-dev/A6xFhcPKdws
(see post of June 23, 2013)